Turn on OSS only exp flag for mfT users too
