+++ mirrord/layer/src/lib.rs
@@ -240,7 +240,6 @@ fn layer_pre_initialization() -> Result<(), LayerError> {
                 sip_binaries_dir: config
                     .experimental
                     .sip_utils
-                    .unwrap_or_default()
                     .then(|| mirrord_sip::MIRRORD_BINARIES_DIR_PATH_BUF.as_path()),
             },
             log_info,
